Refactor project structure, grouping utilities.
